add some basic comments
